+++ activemodel/test/cases/type/binary_test.rb
@@ -11,6 +11,12 @@ module ActiveModel
         assert_equal "1", type.cast("1")
         assert_equal 1, type.cast(1)
       end
+
+      def test_serialize_binary_strings
+        type = Type::Binary.new
+        assert_equal "ƒée".b, type.serialize("ƒée")
+        assert_not_equal "ƒée", type.serialize("ƒée")
+      end
     end
   end
 end